Add files via upload
I first chose a data set within the COVID data that didn't have enough variation to be interesting, at least in its cumulative state. So then I settled on studying the daily percentage increase of new COVID cases in Utah.  However, as I worked through the steps of the project, the smoothing went well and everything was great until I had to fit an AR model. 

I learned a lot about different data sets and how some are not suitable for regression analysis for different reasons-- those things I included in the attached report. 

By picking a data set that should and does peter out with the eradicating of the virus that the data is generated from, I cannot expect necessarily to be able to forecast anything groundbreaking, and a regression model might not be the best fit at all.

This project helped me dive deeper into the nature of regression analysis and I learned more about what we can learn from residual plots, etc. 
At any rate, the calculation of Utah's Immune population percentage that I calculated at the end is valid, interesting, and fun.